ADD sistema de permisos en creacion y elminacion de productos de la base de datos

+++ src/routes/api/mongooseProduct.router.js
@@ -2,6 +2,7 @@ import { Router } from 'express'
 import DaoProductsManager from '../../dao/mongo/managers/products.js'
 import uploader from '../../services/uploader.js'
 import { buildResponsePaginated } from '../../utils.js'
+import { passportCall } from '../../services/auth.js'
 
 import ProductModel from '../../dao/mongo/models/product.js'
 
@@ -41,34 +42,44 @@ router.get('/:pid', async (req, res) => {
   }
 })
 
-router.post('/', uploader.array('thumbnail'), async (req, res) => {
-  const { title, category, description, price, code, stock, status } = req.body
-  if (!title || !category || !description || !price || !code || !stock) {
-    return res
-      .status(400)
-      .send({ status: 'error', error: 'Incomplete values' })
-  }
+router.post('/', passportCall('jwt', { strategyType: 'jwt' }), uploader.array('thumbnail'), async (req, res) => {
+  const userRole = req.user.role
+  if (userRole === undefined) return res.status(401).send({ status: 'error', error: 'Not authenticated' })
+  const authorizedRole = 'premium'
+  if (userRole !== authorizedRole) return res.status(403).send({ status: 'error', error: 'Forbidden' })
+  try {
+    const owner = req.user.id
+    const { title, category, description, price, code, stock, status } = req.body
+    if (!title || !category || !description || !price || !code || !stock) {
+      return res
+        .status(400)
+        .send({ status: 'error', error: 'Incomplete values' })
+    }
 
-  const newProduct = {
-    title,
-    category,
-    description,
-    price,
-    code,
-    stock,
-    status
-  }
-  const thumbnail = req.files.map(
-    (file) =>
-      `${req.protocol}://${req.hostname}:${process.env.PORT || 8080}/img/${file.filename}`
+    const newProduct = {
+      title,
+      category,
+      description,
+      price,
+      code,
+      stock,
+      status,
+      owner
+    }
+    const thumbnail = req.files.map(
+      (file) =>
+        `${req.protocol}://${req.hostname}:${process.env.PORT || 8080}/img/${file.filename}`
 
-  )
-  if (!thumbnail) {
-    newProduct.thumbnail = []
+    )
+    if (!thumbnail) {
+      newProduct.thumbnail = []
+    }
+    newProduct.thumbnail = thumbnail
+    const result = await productsService.createProduct(newProduct)
+    res.send({ status: 'success', payload: result._id })
+  } catch (err) {
+
   }
-  newProduct.thumbnail = thumbnail
-  const result = await productsService.createProduct(newProduct)
-  res.send({ status: 'success', payload: result._id })
 })
 
 router.put('/:pid', async (req, res) => {
@@ -95,11 +106,20 @@ router.put('/:pid', async (req, res) => {
   res.send({ status: 'success', message: 'Product updated' })
 })
 
-router.delete('/:pid', async (req, res) => {
-  const pid = req.params.pid
-  const result = await productsService.deleteProduct(pid)
-  console.log(result)
-  res.send({ status: 'success', message: 'Product deleted' })
+router.delete('/:pid', passportCall('jwt', { strategyType: 'jwt' }), async (req, res) => {
+  const userRole = req.user.role
+  if (userRole === undefined) return res.status(401).send({ error: 'Not authenticated' })
+  const authorizedRoles = ['admin', 'premium']
+  if (!userRole.includes(authorizedRoles)) return res.status(403).send({ status: 'error', error: 'Forbidden' })
+  try {
+    const pid = req.params.pid
+    const result = await productsService.deleteProduct(pid)
+    console.log(result)
+    res.send({ status: 'success', message: 'Product deleted' })
+  } catch (err) {
+    req.logger.error(err)
+    return res.status(500).json({ error: 'An unexpected error ocurred' })
+  }
 })
 
 export default router
